fix: allocates a function for hover to module
