+++ rs-css/src/scripts/app.js
@@ -1,4 +1,5 @@
 import createLevels from './task_template.js';
+import { hovered, unhovered } from './setHoveredElements.js';
 
 const cssInput = document.querySelector('input');
 const enterBtn = document.querySelector('#enter');
@@ -22,7 +23,6 @@ function win() {
 }
 
 function makeAGuess() {
-  // достать селектор из инпута
   let selector = cssInput.value;
   try {
     // наводится на все эл. внутри table и ищет там selector
@@ -33,7 +33,6 @@ function makeAGuess() {
   }
   // console.log(selector);
   // console.log(guessEls); // возвращает NodeList с совпадениями
-  // добавить элементу/элементам класс 'selected'
   // учесть, что может придти одна нода или их array
   // учесть ситуацию, когда по предположенному селектору нет Node
   for (let i = 0; i < guessEls.length; i++) {
@@ -43,12 +42,13 @@ function makeAGuess() {
   return guessEls;
 }
 
+// добавить сброс состояния после win() / fail()
 function checkAnswer() {
   // достать из объекта текущего уровня селектор-ответ
   let correctSelector = levels[currentLevel]
-  console.log('answer: ' + correctSelector.answer);
+  // console.log('answer: ' + correctSelector.answer);
   let correctEls = table.querySelectorAll(correctSelector.answer);  // Nodelist
-  console.log(correctEls[0].classList)
+  // console.log(correctEls[0].classList)
   // let guessEls = makeAGuess();  // возвращает NodeList с совпадениями
   let result;
 
@@ -62,16 +62,13 @@ function checkAnswer() {
       result = true;
     }
   }
+  
   if (result === true) {
     win()
   } else if (result === false) {
     fail()
   }
 }
-
-
-// если checkAnswer() return true то запускаем win()
-// если checkAnswer() return false то запускаем fail()
 
 
 let table = document.querySelector('#table');
@@ -87,72 +84,11 @@ let currentLevel = 3;
 table.innerHTML = levels[currentLevel].divTemplate;
 markup.innerHTML = levels[currentLevel].markupTemplate;
 
-
-// функция setHoveredElements() - берет селектор, на котором мышь(parent)
-//вставляет его в querySelector
-// применять этот querySelector к table и markup одновременно
-// на выбранные элементы вешает класс hover
-// let selectedEl;
-function setHoveredElements(parent, needUnhover) {
-  if(parent.hasChildNodes()) {
-    let children = parent.children;
-    let mirrorChildren;
-    if (parent === table) {
-      mirrorChildren = markup.children;
-      // console.log(parent, mirrorChildren)
-    } else if (parent === markup) {
-      mirrorChildren = table.children;
-      // console.log(parent)
-    }
-    if(needUnhover === 'yes') {
-      for (let i = 0; i < children.length; i++) {
-          mirrorChildren[i].classList.remove('hover');
-      }
-    } else if (needUnhover === 'no') {
-      for (let i = 0; i < children.length; i++) {
-        if (children[i].classList.contains('hover')) {
-          mirrorChildren[i].classList.add('hover');
-        }
-      }
-    }
-    // console.log(children);
-    // console.log(children[0].compareDocumentPosition(mirrorChildren[0]))
-  } else {
-    console.log('no children');
-  }
-}
-
-// setHoveredElements(table)
-
 markup.addEventListener('mouseover', hovered)
 markup.addEventListener('mouseout', unhovered)
 table.addEventListener('mouseover', hovered)
 table.addEventListener('mouseout', unhovered)
 
-function unhovered(e) {
-  if (e.target.id == 'table' || e.target.id == 'markup') { return }
-  // e.target.classList.remove('hover');
-  if (e.target.parentNode.id === 'table') {
-    e.target.classList.remove('hover');
-    setHoveredElements(table, 'yes');
-  } else if (e.target.parentNode.id === 'markup') {
-    e.target.classList.remove('hover');
-    setHoveredElements(markup, 'yes');
-  }
-}
-function hovered(e) {
-  if (e.target.id == 'table' || e.target.id == 'markup') { return }
-  // console.log(e.target.parentNode.id)
-  if (e.target.parentNode.id === 'table') {
-    e.target.classList.add('hover');
-    setHoveredElements(table, 'no');
-  } else if (e.target.parentNode.id === 'markup') {
-    e.target.classList.add('hover');
-    setHoveredElements(markup, 'no');
-  }
-  // console.log('ты трогал меня - ' +  e.target.tagName.toLowerCase(), e.target.classList.value);
-  e.target.classList.add('hover');
-}
-
 enterBtn.addEventListener('click', checkAnswer);
 
+export { table, markup }
